+++ pkg/domain/blueprintSpec.go
@@ -275,6 +275,9 @@ func (spec *BlueprintSpec) MissingConfigReferences(error error) {
 func (spec *BlueprintSpec) DetermineStateDiff(
 	ecosystemState ecosystem.EcosystemState,
 	referencedSensitiveConfig map[common.DoguConfigKey]common.SensitiveDoguConfigValue,
+	referencedConfig map[common.DoguConfigKey]common.DoguConfigValue,
+	referencedSensitiveGlobalConfig map[common.GlobalConfigKey]common.GlobalConfigValue,
+	referencedGlobalConfig map[common.GlobalConfigKey]common.GlobalConfigValue,
 	isDebugModeActive bool,
 ) error {
 	doguDiffs := determineDoguDiffs(spec.EffectiveBlueprint.Dogus, ecosystemState.InstalledDogus)
@@ -288,6 +291,9 @@ func (spec *BlueprintSpec) DetermineStateDiff(
 		ecosystemState.ConfigByDogu,
 		ecosystemState.SensitiveConfigByDogu,
 		referencedSensitiveConfig,
+		referencedConfig,
+		referencedSensitiveGlobalConfig,
+		referencedGlobalConfig,
 	)
 
 	spec.StateDiff = StateDiff{