+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#129] Reconciliation of the blueprint on changes of dogu-crs, ces-configMaps and ces-secrets
 - [#131] Ignore loglevel changes while debug-mode is active
 - [#131] Do not reconcile blueprint if a restore is in progress
+- [#136] Support non-sensitive config references with config maps and secrets
 
 ### Changed
 - [#119] *breaking* sensitive dogu config can now only be referenced with secrets
@@ -30,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#121] Upgrade Makefiles to v10.4.0
 - [#121] *breaking* merge proxy config dogu action into one to simplify the status
 - [#133] updated and added docs for the operator
+- [#136] Prevent sensitive configs in global config
 
 ### Removed
 - [#119] *breaking* no support for v1 blueprint CRs anymore
@@ -42,6 +44,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#121] *breaking* remove the ability to apply components (including self upgrade)
   - Ecosystem-Core-Chart has now the resonsibility of components
 - [#121] *breaking* remove watching the health state of components in the ecosystem
+
+### Fixed
+- [#141] fix nilpointer with missing blueprint mask
 
 ## [v2.8.0] - 2025-09-15
 ### Changed